requirements/dev.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.26.post2
celery-with-redis==3.0
Django==2.0.3
django-celery==3.2.2
django-ckeditor==5.5.0
django-js-asset==1.1.0
django-ranged-response==0.2.0
django-redis==4.9.0
django-simple-captcha==0.5.7
django-mptt==0.10.0
jieba==0.39
kombu==3.0.37
Pillow==5.1.0
PyMySQL==0.8.1
pytz==2018.4
redis==2.10.6
six==1.11.0
